pyproject.toml
[project]
name = "library-fastapi"
version = "1.0.0"
description = "Example project for FastAPI and uv"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "psycopg>=3.2.7",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.1", # not maintained
    "httpx>=0.28.1", # FastAPI TestClient requires it
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
]
lint = [
    "ruff>=0.11.8",
]

[tool.pytest.ini_options]
env = [
    "db_host=localhost",
    "db_port=15432",
    "db_name=library",
    "db_user=user",
    "db_password=password"
]

[tool.ruff]
line-length = 100
target-version = "py313"
